requirements.txt
# Kayan SDG Data Pipeline — Requirements
llama-cloud>=2.1
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
python-pptx>=0.6.21
chardet>=5.0.0
tqdm>=4.65.0
httpx>=0.24.0
tiktoken>=0.5.0
